Solution to Logical OR Operation

+++ Logical OR Operation@@67882a731b4392d8fc1577dd/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a,b;
-    scanf("%d%d",&a,&b);
-    if(a||b >0)
+    float a,b;
+    scanf("%f%f",&a,&b);
+    if(a||b>0)
     {
         printf("True");
     }
